hardware controller tweaks from PR comments
